Use f64 for frequencies Not strictly necessary, but it matches the JS version

+++ src/index.rs
@@ -55,7 +55,7 @@ impl Index {
             }
 
             for (token, count) in &token_freq {
-                let freq = (*count as f32).sqrt();
+                let freq = (*count as f64).sqrt();
                 self.index.get_mut(field)
                     .expect("Invalid HashMap") // TODO: better API
                     .add_token(doc_ref, token, freq);
